feat(db): implement MongoDB repository for Game entity with reactive support
