+++ src/main/java/com/itacademy/blackjack/game/infrastructure/persistence/mongo/repository/GameRepository.java
@@ -1,4 +1,4 @@
-package com.itacademy.blackjack.game.domain.repository;
+package com.itacademy.blackjack.game.infrastructure.persistence.mongo.repository;
 
 import com.itacademy.blackjack.game.domain.model.Game;
 import reactor.core.publisher.Mono;
@@ -8,5 +8,5 @@ import java.util.UUID;
 public interface GameRepository {
     Mono<Game> save(Game game);
     Mono<Game> findById(UUID id);
-    Mono<Void> delete(UUID id);
+    Mono<Void> deleteById(UUID id);
 }
