+++ src/main/java/com/itacademy/blackjack/game/infrastructure/persistence/mongo/repository/GameRepositoryImpl.java
@@ -1,0 +1,36 @@
+package com.itacademy.blackjack.game.infrastructure.persistence.mongo.repository;
+
+import com.itacademy.blackjack.game.domain.model.Game;
+import com.itacademy.blackjack.game.infrastructure.persistence.mongo.document.GameDocument;
+import com.itacademy.blackjack.game.infrastructure.persistence.mongo.mapper.GameMapper;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+import reactor.core.publisher.Mono;
+
+import java.util.UUID;
+
+@Repository
+@RequiredArgsConstructor
+public class GameRepositoryImpl implements GameRepository {
+
+    private final GameMongoRepository mongoRepository;
+    private final GameMapper mapper;
+
+    @Override
+    public Mono<Game> save(Game game) {
+        GameDocument document = mapper.toDocument(game);
+        return mongoRepository.save(document)
+                .map(mapper::toDomain);
+    }
+
+    @Override
+    public Mono<Game> findById(UUID id) {
+        return mongoRepository.findById(id.toString())
+                .map(mapper::toDomain);
+    }
+
+    @Override
+    public Mono<Void> deleteById(UUID id) {
+        return mongoRepository.deleteById(id.toString());
+    }
+}
